requirements.txt
Pillow==10.0.1
py2app==0.28.7
